+++ src/protect/strenc/codeenc.rs
@@ -16,14 +16,14 @@ pub fn region(img: &PeImage) -> Option<(u32, usize, u32)> {
     None
 }
 
-pub fn prepare(img: &mut PeImage, encryptCode: bool, key: u32) -> (u32, u32, Vec<u8>) {
+pub fn prepare(img: &mut PeImage, encryptCode: bool, key: u64) -> (u32, u32, Vec<u8>) {
     match (encryptCode, region(img)) {
         (true, Some((rva, _, len))) => (rva, len, encrypt(img, key)),
         _ => (0, 0, Vec::new()),
     }
 }
 
-pub fn encrypt(img: &mut PeImage, key: u32) -> Vec<u8> {
+pub fn encrypt(img: &mut PeImage, key: u64) -> Vec<u8> {
     if let Some((_, raw, len)) = region(img) {
         let end = (raw + len as usize).min(img.data.len());
         cipher::stream(&mut img.data[raw..end], key, 0);
